feat(UNT-T33782): feedback changes and enable clear sync data option
